split readline into it's own module directory
